+++ tests/testthat/test-boost_tree_xgboost.R
@@ -939,7 +939,7 @@ test_that('interface to param arguments', {
 
 test_that('xgboost execution, quantile regression', {
   skip_if(getRversion() <= "4.2.3")
-  skip_if_not_installed("xgboost")
+  skip_if_not_installed("xgboost", minimum_version = "3.4.0.0")
   skip_if_not_installed("modeldata")
   skip_on_cran()